Revert "Pulling forked branch for develop testing  (#109)" (#110)
This reverts commit d70175dabd47140139e7a2a8627ecffc92923723.
