--- docs/Gemfile
@@ -1,15 +1,0 @@
-source "https://rubygems.org"
-
-# Use remote-theme in _config.yml
-# gem "just-the-docs", "~> 0.3.0"
-gem "github-pages", group: :jekyll_plugins
-group :jekyll_plugins do
-  gem "jekyll-feed", "~> 0.12"
-end
-
-
-platforms :mingw, :x64_mingw, :mswin, :jruby do
-  gem "tzinfo", "~> 1.2"
-  gem "tzinfo-data"
-end
-gem "wdm", "~> 0.1.1", :platforms => [:mingw, :x64_mingw, :mswin]
