--- docs/_config.yml
@@ -1,4 +1,0 @@
-remote_theme: pmarsceill/just-the-docs
-baseurl: /tranql
-
-title: TranQL
