--- docs/tranql-jupyter-example-notebook.md
@@ -1,30 +1,0 @@
----
-layout: default
-title: Example Notebook
-nav_order: 1
-parent: TranQL Jupyter
----
-# Example Notebook
-<style>
-html, body, #top {
-  height: 100%;
-}
-#top {
-  display: flex;
-  flex-direction: column;
-}
-#main-content-wrap {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-}
-#main-content {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-}
-iframe {
-  flex-grow: 1;
-}
-</style>
-<iframe frameborder="0" src="https://nbviewer.jupyter.org/github/frostyfan109/TranQL-Jupyter/blob/master/test_notebooks/Demo%20Notebook.ipynb"></iframe>
