+++ tranql/grammar.py
@@ -2,7 +2,7 @@ from pyparsing import (
     Combine, Word, White, Literal, delimitedList, Optional, Empty, Suppress,
     Group, alphas, alphanums, printables, Forward, oneOf, quotedString,
     ZeroOrMore, restOfLine, CaselessKeyword, ParserElement, LineEnd,
-    removeQuotes, Regex, nestedExpr, pyparsing_common as ppc)
+    removeQuotes, Regex, pyparsing_common as ppc)
 
 """
 A program is a list of statements.
@@ -47,35 +47,13 @@ binop = oneOf("= != =~ !=~ < > >= <= eq ne lt le gt ge", caseless=True)
 realNum = ppc.real()
 intNum = ppc.signed_integer()
 
-function_body = Forward()
-# Valid data types: nested function, TranQL variable, real, integer, string
-# For a normal arg, this is the entire argument. For a named arg, this is the actual value.
-arg_value = function_body | ident | realNum | intNum | quotedString
-# Named arguments can be in a Group, so that they are a list in the ast (e.g. `named_arg1=23` => `['named_arg1', '=', 23]`)
-# Since lists are not a supported type any list can be deduced to be a named argument
-# For some reason by adding a parse action to function_body, named_arg has to be manually converted from a Group to a list
-named_arg = Group(Word(alphanums+"_") + "=" + arg_value).setParseAction(lambda t: t.asList())
-
-function_arg = named_arg | arg_value
-
-function_body <<= Word(alphanums+"_") + (
-    Literal("(").suppress() + (delimitedList(function_arg) | Empty()) + Literal(")").suppress()
-)
-# Since asList is called in the TranQL ast, a function ends up being structured as ["my_function_name", ["my_arg1", "my_arg"] or ["add_int", [4, 7]]
-# Accordingly, there is no way to distinguish a function from an actual list. Since asList is called, we cannot give function_body a name.
-# Therefore, the most straightforward way is to set a parsing action which converts the function structure to an actual class.
-# However, classes are not easily json serializable, so a dict struct will do
-function_body.setParseAction(lambda toks: { "name" : toks[0], "args" : toks[1:] })
-
-concept_or_var_list = Group(LBRACK.suppress() + delimitedList(concept_value | ident) + RBRACK.suppress())
 # need to add support for alg expressions
-columnRval = function_body | realNum | intNum | quotedString.addParseAction(removeQuotes) | columnName | concept_or_var_list
+columnRval = realNum | intNum | quotedString.addParseAction(removeQuotes) | columnName | concept_value_list
 whereCondition = Group(
-    ( columnName + binop + columnRval ) |
-    # ( columnName + in_ + concept_value_list) |
-    # ( columnName + in_ + "(" + delimitedList( columnRval ) + ")" ) |
-    # ( columnName + in_ + "(" + statement + ")" ) |
-    # ( columnName + in_ + (columnRval | Word(printables))) |
+    ( columnName + binop + (columnRval | Word(printables) ) ) |
+    ( columnName + in_ + "(" + delimitedList( columnRval ) + ")" ) |
+    ( columnName + in_ + "(" + statement + ")" ) |
+    ( columnName + in_ + (columnRval | Word(printables))) |
     ( "(" + whereExpression + ")" )
 )
 whereExpression << whereCondition + ZeroOrMore( ( and_ | or_ ) + whereExpression )
