--- tranql/udfs.yaml
@@ -1,16 +1,0 @@
----
-  userDefinedFunctions:
-    modules:
-      - source: udfs.py
-        functions:
-          - children
-          - descendants
-          - ancestors
-          - parents
-          - siblings
-          - close_match
-          - exact_match
-
-          - AND
-          - OR
-          - XOR
