+++ web/package.json
@@ -7,12 +7,9 @@
     "@material-ui/lab": "^3.0.0-alpha.30",
     "abortcontroller-polyfill": "^1.3.0",
     "bootstrap": "^4.3.1",
-    "classnames": "^2.2.6",
     "create-react-class": "^15.6.3",
     "dexie": "^2.0.4",
     "dexie-observable": "^1.0.0-beta.5",
-    "elasticlunr": "^0.9.5",
-    "es-to-primitive": "^1.2.1",
     "file-saver": "^2.0.2",
     "filepond": "^4.4.11",
     "filepond-plugin-file-validate-type": "^1.2.4",
