+++ web/src/App.css
@@ -88,7 +88,6 @@ body {
   /* color:black; */
 }
 #graphOverlayContainer {
-  position:relative;
   flex-grow:1;
   height:100%;
   display:flex;
@@ -123,10 +122,8 @@ body {
   max-height:400px;
 }
 .FindTool {
-  position: absolute;
-  width: calc(100% - 16px);
   margin: 8px;
-  /* margin-left:0; */
+  margin-left:0;
   /* 16px for margin */
   max-height:calc(100% - 16px);
   /* position:absolute;
@@ -158,7 +155,7 @@ body {
 }
 .tableViewPane {
   overflow-y:auto;
-  /* padding:15px; */
+  padding:15px;
 }
 #info {
     position: relative;
@@ -317,8 +314,9 @@ body {
 .import-export-modal-dialog .modal-body {
   display:flex;
 }
-.import-export-modal-dialog .modal-body .no-select {
+.import-export-modal-dialog .modal-body > * {
   /* flex:1; */
+  margin:0 5px;
   background-color:rgb(245,245,245);
   border-radius:5px;
   border:1.5px solid rgb(230,230,230);
@@ -327,21 +325,11 @@ body {
   flex-direction:column;
   min-height:200px;
 }
-.import-export-modal-dialog .modal-body .row > * {
-  margin:0 5px;
-}
-.import-export-modal-dialog .modal-body .no-select > * {
+.import-export-modal-dialog .modal-body > * > * {
   padding:10px;
 }
-.import-export-modal-dialog .modal-body .no-select > *:not(:last-child) {
+.import-export-modal-dialog .modal-body > * > *:not(:last-child) {
   padding-bottom:0;
-}
-.import-export-modal-dialog .copy-url-container {
-  flex-grow: 1;
-  margin-top: 10px !important;
-}
-.import-export-modal-dialog .copy-url-container input {
-  background-color: initial;
 }
 .import-export-icon-container {
   width:100%;
@@ -424,8 +412,7 @@ body {
   flex-direction:column;
 }
 .toolbar-help-modal-dialog .modal-body > .tab-content {
-  height:0; /* ?? */
-  flex-grow:1; /* ?? the height:0 is required to fix overflow for some reason */
+  height:calc(90% - 5px);
   margin-top:10px;
   display:flex;
 }
