+++ web/src/App.js
@@ -2,19 +2,21 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { css } from '@emotion/core';
 import { Button } from 'reactstrap';
-import { Modal, Form, Card, Container, Row, Col, ListGroup, Table, Tabs, Tab, InputGroup } from 'react-bootstrap';
+import { Modal, Form, Card, Container, Row, Col, ListGroup, Table, Tabs, Tab } from 'react-bootstrap';
 import { ForceGraph3D, ForceGraph2D, ForceGraphVR } from 'react-force-graph';
 import * as sizeof from 'object-sizeof';
 import JSONTree from 'react-json-tree';
 import * as JSON5 from 'json5';
+import * as YAML from 'js-yaml';
 import * as qs from 'qs';
+import FileSaver from 'file-saver';
 // import logo from './static/images/tranql.png'; // Tell Webpack this JS file uses this image
 import { contextMenu } from 'react-contexify';
 import { IoIosArrowDropupCircle, IoIosArrowDropdownCircle, IoIosSwap, IoMdBrowsers } from 'react-icons/io';
 import {
   FaCog, FaDatabase, FaQuestionCircle, FaSearch, FaHighlighter, FaEye,
   FaSpinner, FaMousePointer, FaTimes, FaFolderOpen, FaFileImport, FaFileExport,
-  FaArrowsAlt, FaTrash, FaPlayCircle, FaTable, FaCopy, FaPython
+  FaArrowsAlt, FaTrash, FaPlayCircle, FaTable
 } from 'react-icons/fa';
 // import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
 import 'react-confirm-alert/src/react-confirm-alert.css';
@@ -29,21 +31,18 @@ import { Range } from 'rc-slider';
 import { GridLoader } from 'react-spinners';
 import SplitPane from 'react-split-pane';
 import Cache from './Cache.js';
+import FileLoader from './FileLoader.js';
 import AnswerViewer from './AnswerViewer.js';
 import QueriesModal from './QueriesModal.js';
-import HistoryViewer from './HistoryViewer.js';
 import BrowseNodeInterface from './BrowseNodeInterface.js';
 import Legend from './Legend.js';
 import TableViewer from './TableViewer.js';
-import HelpModal, { ToolbarHelpModal } from './HelpModal.js';
-import ImportExportModal from './ImportExportModal.js';
 import confirmAlert from './confirmAlert.js';
 import highlightTypes from './highlightTypes.js';
-import { shadeColor, adjustTitle, hydrateState, formatBytes } from './Util.js';
+import { shadeColor, adjustTitle, scrollIntoView, hydrateState, formatBytes } from './Util.js';
 import { Toolbar, Tool, /*ToolGroup*/ } from './Toolbar.js';
 import LinkExaminer from './LinkExaminer.js';
-// import FindTool from './FindTool.js';
-import FindTool2 from './FindTool2.js';
+import FindTool from './FindTool.js';
 import Message from './Message.js';
 import Chain from './Chain.js';
 import ContextMenu from './ContextMenu.js';
@@ -163,7 +162,12 @@ class App extends Component {
     // Type chart
     this._renderTypeChart = this._renderTypeChart.bind (this);
 
-    this._setActiveModal = this._setActiveModal.bind (this);
+    // Help modal
+    this._renderHelpModal = this._renderHelpModal.bind (this);
+    this._renderToolbarHelpModal = this._renderToolbarHelpModal.bind (this);
+
+    // Import/Export modal
+    this._renderImportExportModal = this._renderImportExportModal.bind (this);
 
     // Annotate graph
     this._annotateGraph = this._annotateGraph.bind (this);
@@ -172,11 +176,12 @@ class App extends Component {
     this._browseNodeResult = this._browseNodeResult.bind (this);
 
     // Settings management
+    this._handleShowModal = this._handleShowModal.bind (this);
+    this._handleCloseModal = this._handleCloseModal.bind (this);
     this._handleUpdateSettings = this._handleUpdateSettings.bind (this);
     this._toggleCheckbox = this._toggleCheckbox.bind (this);
     this._renderCheckboxes = this._renderCheckboxes.bind (this);
     this._hydrateState = hydrateState.bind (this);
-    this._handleQueryString = this._handleQueryString.bind (this);
 
     this._handleShowAnswerViewer = this._handleShowAnswerViewer.bind (this);
     this._handleMessageDialog = this._handleMessageDialog.bind (this);
@@ -205,6 +210,9 @@ class App extends Component {
     this._exampleQueriesModal = React.createRef ();
     this._cachedQueriesModal = React.createRef ();
 
+    // Import/Export modal
+    this._importForm = React.createRef ();
+    this._exportForm = React.createRef ();
 
     // Create the graph's GUI-related references
     this._graphSplitPane = React.createRef ();
@@ -310,11 +318,6 @@ class App extends Component {
       tableView : false,
       tableViewerSize : 1 - (2/7),
 
-      // Keep track of table viewer/interactive shell plugins
-      tableViewerComponents : {
-        tableViewerCompActive : false,
-      },
-
       // Schema viewer
       schema : {
         nodes : [],
@@ -346,9 +349,289 @@ class App extends Component {
       // Tools for the toolbar component
       useToolCursor : false,
 
+      toolHelpDescriptions : {
+        tools : [
+          {
+            title: "Navigate",
+            description: `
+            This tool is intended to make navigating the force graph more easy to do. Left clicking a node will pan the camera to it and zoom in on it.
+            It will also make it the center of rotation for the camera.
+            Left clicking and holding a node will not trigger this tool and will instead drag the node like normal.
+            Right clicking and dragging will shift the center of rotation away from the node selected with this tool.`
+          },
+          {
+            title: "Select",
+            description: `
+            This tool allows you to view the additional data that nodes and links may possess. To use it, left click a node or link, which will bring up the object viewer.
+            From there, you can navigate the tree view of the selected object and view all the properties that it has. The object viewer is inside a horizontal split pane
+            along with the graph, so it may be resized by dragging the border where the two intersect.`
+          },
+          {
+            title: "Highlight Types",
+            description: `
+            This tool allows you to highlight all the nodes or links that share a type. It will highlight the type of the node or link that the cursor is currently hovering over.
+            For nodes and links of multiple types, it will highlight all other nodes or links that share any types with it.
+            Left clicking a node or link with this tool will hide all highlighted elements. Right clicking a node or link with this tool will hide all non-highlighted elements.`
+          },
+          {
+            title: "Examine Connection",
+            description: `
+            This tool allows you to view all links, and the direction of each link, existing betweeen two nodes.
+            To use the tool, left click any link between a pair of desired nodes.
+            This will bring up the connection viewer interface.
+            It will display each node's name as an abbreviation. It also colors codes their names according to the node's color in the force graph.
+            If you forget a node's name, you can hover over the abbreviation, and it will display its name in full.
+            Clicking on a link will bring it up in the object viewer.
+            `
+          },
+          {
+            title: "Browse Node",
+            description: `
+            This tool allows you to select a node and browse connections it has with nodes of a given type, optionally along a given prediate. It will
+            query the Robokop API and return new nodes.
+            `
+          }
+        ],
+        buttons: [
+          {
+            title: "Answer Navigator",
+            description: "This button brings up Robokop's depth analysis of the answer. It also displays a variety of other data related to the graph."
+          },
+          {
+            title: "Find Tool",
+            description: (
+              <div>
+                <Tabs className="find-tool-tabs" defaultActiveKey="overview">
+                  <Tab eventKey="overview" title="Overview">
+                    <p>
+                      This button brings up the find tool, which can also be opened with the keyboard shortcut control+F (the normal browser find tool can be opened with F3).
+                      The find tool enables you to use JSONPath or JSON-like attribute selectors to find objects in the graph. Additionally, you may use tools on the results.
+                      <br/><br/>
+                      The JSONPath syntax is limited relative to the normal syntax in that you cannot select node pairs ({`nodes->links->nodes`}).
+                      However, JSONPath also allows for the exploration of the entire graph object, rather than only selecting nodes and links.
+                      Another advantage of the JSONPath syntax is that you do not need to know JSONPath to use it, as you can explore the graph via the arrows
+                      on the results.
+                    </p>
+                  </Tab>
+                  <Tab eventKey="normalSyntax" title="Normal Syntax">
+                    <div className="section">
+                      <h6>Structure:</h6>
+                      <p>
+                        The general structure of queries is `selector`{"{`attributes`}"}. A selector can be either "nodes", "links", or "*". The asterisk selector selects both nodes and links.
+                        You can also connect these selectors with transitions, with the structure `selector`{"{`attributes`}"} -> `selector`{"{`attributes`}"} -> `selector`{"{`attributes`}"}. Note: only links are applicable in the second selector.
+                        The `{"{}"}` following the selector may be omitted if no attributes are present.
+                        Attributes must be valid <a target="_blank" rel="noopener noreferrer" href="https://tools.ietf.org/html/rfc7159">JSON</a>. This means that attributes are structured as key to value, where key is an attribute that a node or link may or may not possess.
+                        If you are unsure as to what attributes nodes and links have, you can use the <FaMousePointer style={{fontSize:"14px"}}/> select tool to view a node or link's attributes.
+                        However, keep in mind, only some attributes such as "id", "name", "type", and "equivalent_identifiers" are standard. Not all nodes or links are gaurenteeed to have others.
+                      </p>
+                      <div className="section">
+                        <h6>Flags:</h6>
+                        <p>
+                          A colon in an attribute key indicates that the following text is an attribute flag.
+                          Attribute flags modify the behavior of how the attribute is compared to the provided value.
+                          For example, if an attribute is a list (['a','b','c']), you can use the `includes` (pseudo) flag to check if `a` is in the list.
+                          This would look like "nodes{`{"attribute:includes" : "a"}`}".
+                          All normal colons inside of attribute keys must be escaped, i.e. preceded by a backslash ("\:"),
+                          or it will be assumed that the following text is an attribute flag. A scenario in which it would be necessary to do this would be
+                          nodes{`{"a:test":'value'}`} where you wanted to select nodes whose attribute "a:test" has the value "value".
+                          Only one attribute flag can currently be used in an attribute.<br/>
+                          <span style={{fontWeight:600}}>Example:</span> nodes{`{"omnicorp_article_count:>=":100}`} selects all nodes whose `omnicorp_article_count`
+                          attribute is greater than or equal to 100. The flag in this query is ">=".
+                        </p>
+                        <div className="section">
+                          <h6>The valid flags are:</h6>
+                          <Row>
+                            <dl>
+                              {/*eslint-disable-next-line*/}
+                              <Col><dt>regex</dt></Col><Col><dd>Matches a <a target="_blank" rel="noopener noreferrer" href="http://cecas.clemson.edu/~warner/M865/RegexBasics.html">regular expression</a> against the element's attribute<br/><a href="javascript:void(0)" onClick={
+                                ()=>scrollIntoView("#regexFlag")
+                              }>Example</a></dd></Col>
+                              {/*eslint-disable-next-line*/}
+                              <Col><dt>func</dt></Col><Col><dd>Evals a JavaScript function which is passed the element's attribute as the only argument. Should return true or false to indicate if the node should or should not be included.<br/><a href="javascript:void(0)" onClick={
+                                ()=>scrollIntoView("#funcFlag")
+                              }>Example</a></dd></Col>
+                              <Col><dt>Comparison Operators</dt></Col><Col><dd>The comparison operators <b>({"<"}, {"<="}, {">"}, {">="}, {"=="}, {"==="}, {"!="}, {"!=="})</b> are all valid flags. {"<="} and {">="} compare if a number is less than or equal to and greater than or equal to the given input respectively.
+                              "!=" and "!==" both compare if it is not equal to the given input. The "===" and "!==" operators do not allow for implicit conversion when comparing, and therefore should almost always be used over their alternative.</dd></Col>
+                              <Col>
+                                <dt>Pseudo Flags</dt>
+                                <dd>
+                                  Any other method or property in the element's attribute's JavaScript prototype chain (for common references see&nbsp;
+                                  <a target="_blank" rel="noopener noreferrer" href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/prototype">Text</a> and&nbsp;
+                                  <a target="_blank" rel="noopener noreferrer" href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/prototype">Lists</a>).
+                                  {/*eslint-disable-next-line*/}
+                                  <br/><a href="javascript:void(0)" onClick={
+                                    ()=>scrollIntoView("#regexFlag")
+                                  }>Example (list includes)</a>
+                                </dd>
+                              </Col>
+                            </dl>
+                          </Row>
+                        </div>
+                      </div>
+                      <div className="section">
+                        <h6>Magic variables (no current uses)</h6>
+                        <p>
+                          While flags are concerning attribute keys, magic variables are used in attribute values.
+                          These variables use the syntax "__`variable name`__". They are used as if they are plain text, for example, "nodes->links{`{"type:regex":"foo|__sourceNodes__"}`}->nodes".
+                          All instances of a normal two underscores in a row must be escaped with a backslash ("\__"), or it will result in any text following it and preceding another two unescaped underscores being detected as a magic variable.
+                        </p>
+                        <div className="section">
+                          <h6>The valid magic variables are:</h6>
+                          <Row>
+                            <dl>
+                              <Col><dt>__sourceNodes__</dt></Col><Col><dd>(Only applicable in the second selector of a nodes->links->nodes query) The value of __sourceNodes__ is a regex string that matches for any ids of the nodes from the source selector.</dd></Col>
+
+                              <Col><dt>__targetNodes__</dt></Col><Col><dd>(Only applicable in the second selector of a nodes->links->nodes query) The value of __targetNodes__ is a regex string that matches for any ids of the nodes from the target selector.</dd></Col>
+
+                              <Col><dt>__element__</dt></Col><Col><dd>The value of __element__ is the current element being tested for the attribute.</dd></Col>
+
+                              <Col><dt>__nodes__</dt></Col><Col><dd>The value of __nodes__ is a list of the all the nodes in the current force graph</dd></Col>
+
+                              <Col><dt>__links__</dt></Col><Col><dd>The value of __links__ is a list of the all the links in the current force graph</dd></Col>
+                            </dl>
+                          </Row>
+                        </div>
+                      </div>
+                    </div>
+                    <div className="section">
+                      <h6>Example queries:</h6>
+                      <Row>
+                        <dl>
+                          <Col><dt>nodes{`{"id": "chemical_substance"}`}</dt></Col>
+                          <Col><dd>This will find all nodes in the graph (schema) whose `id` attribute equals "chemical_substance".</dd></Col>
+
+                          <Col id="regexFlag"><dt>nodes{`{"equivalent_identifiers:includes":"CHEMBL:CHEMBL3"}`} -><br/> links{`{"type:includes":"related_to"}`} -><br/> nodes{`{"name:regex":"(disease|genetic_condition)"}`}</dt></Col>
+                          <Col><dd>This will find all nodes in the graph who have the curie "CHEMBL:CHEMBL3" in their equivalent_identifiers attribute, all nodes whose name matches the regular expression "(disease|genetic_condition)",
+                          and all links whose source is any node from the first selector, target is from the third selector, and has the type "related_to".</dd></Col>
+
+                          <Col id="funcFlag"><dt>nodes{`{"description:func":"function(description) { return description.split("/").includes('test'); }"}`}</dt></Col>
+                          <Col><dd>This will find all nodes in the graph whose `description` attribute split by a forward slash contains the string "test".</dd></Col>
+                        </dl>
+                      </Row>
+                    </div>
+                  </Tab>
+                  <Tab eventKey="JSONPath" title="JSONPath">
+                    <div className="section">
+                      <h6>Structure:</h6>
+                      <p>
+                        When using JSONPath, the root is the graph object. Every JSONPath result that can be explored
+                        has forward arrows which will set your query to be the result's results. If the results are node or link objects
+                        ($.nodes.{"{}"} or $.links.{"{}"} or $.*.{"{}"}), you can use tools on them like the normal syntax.<br/><br/>
+                        <span style={{fontWeight:"600"}}>If using filters, please note that the object viewer displays the element's `origin` attribute.
+                        The origin attribute contains almost all relevant information.</span><br/><br/>
+                        For an extensive JSONPath reference, see <a target="_blank" rel="noopener noreferrer" href="https://goessner.net/articles/JsonPath/">this</a>.<br/>
+                        For the specific JSONPath module in use, see <a target="_blank" rel="noopener noreferrer" href="https://www.npmjs.com/package/jsonpath">this</a>.
+                      </p>
+                    </div>
+                    <div className="section">
+                      <h6>Example queries:</h6>
+                      <Row>
+                        <dl>
+                          <Col><dt>$.nodes[?(@.origin.id === 'chemical_substance')]</dt></Col>
+                          <Col><dd>This will find all nodes in the graph (schema) whose origin's `id` attribute equals "chemical_substance".</dd></Col>
+
+                          <Col><dt>$.nodes[?(@.origin.equivalent_identifiers.includes("CHEMBL:CHEMBL3"))]</dt></Col>
+                          <Col><dd>This will find all nodes in the graph who have the curie "CHEMBL:CHEMBL3" in their origin's "equivalent_identifiers" attribute.</dd></Col>
+
+                          <Col><dt>$.*[?(@.origin.reasoner.includes('rtx'))]</dt></Col>
+                          <Col><dd>This will find all nodes and links in the graph whose `reasoner` attribute includes "rtx"</dd></Col>
+
+                          <Col><dt>$.links[?(@.origin.source_database.includes('mychem') && @.target.origin.omnicorp_article_count >= 1000)]</dt></Col>
+                          <Col><dd>This will find all links in the graph whose whose origin's source database includes "mychem"
+                          and whose target node's origin has an `omnicorp_article_count` greater than or equal to 1000</dd></Col>
+                        </dl>
+                      </Row>
+                    </div>
+                  </Tab>
+                </Tabs>
+              </div>
+          )
+          },
+          {
+            title: "Help & Information",
+            description: `
+            This button brings up the help and information center, where you can find various references for using TranQL.`
+          },
+          {
+            title: "Cache Viewer",
+            description: `
+            This button brings up the cache viewer interface, which displays all queries which have been locally cached.
+            It allows for you to find previous queries and quickly view them, edit them, or delete them.`
+          },
+          {
+            title: "Import/Export",
+            description: `
+            This button brings up the import/export interface which allows you to import and export TranQL force graphs
+            into various file formats that can then be loaded by others. The graphs retain their visual state upon exportation.`
+          },
+          {
+            title: "Settings",
+            description: (
+              <Tabs defaultActiveKey="overview">
+                <Tab eventKey="overview" title="Overview">
+                  This button brings up the settings interface, which allows you to customize the behavior of TranQL.
+                </Tab>
+                <Tab eventKey="general" title="General">
+                  <h6>Visualization Mode and Graph Colorization</h6>
+                    <p>This allows you to change the way that the graph is visualized. You may also disable the coloring of the graph if desired.</p>
+                  <h6>Use Cache</h6>
+                    <p>This allows you to disable the caching of the TranQL schema and any results from TranQL queries. This means that the results will not be stored locally
+                    on your machine for future use. Disabling the cache will not delete your currently cached queries.
+                    You may also clear the cache if desired.</p>
+                  <h6>Cursor</h6>
+                    <p>This sets your mouse cursor to be the same icon as the currently selected tool.</p>
+                </Tab>
+                <Tab eventKey="graphStructure" title="Graph Structure">
+                  <h6>Link Weight Range</h6>
+                    <p>This will filter out any links from the graph whose weights are not within the specified range.</p>
+                  <h6>Node Connectivity Range</h6>
+                    <p>This will filter out any nodes from the graph whose number of connections (links) with other nodes is not within the specified range.</p>
+                  <h6>Force Graph Charge</h6>
+                    <p>This will set the charge force applied to nodes within the graph. Charge is a property that acts like an electrical charge and causes either
+                    the attraction or repulsion of nodes between one another. It can be used to get a more enlarged and spread out view of the graph and prevent
+                    nodes from overlapping. <a target="_blank" rel="noopener noreferrer" href="https://d3-wiki.readthedocs.io/zh_CN/master/Force-Layout/#charge">More comprehensive reference</a></p>
+                  <h6>Legend Display Limit</h6>
+                    <p>This will filter out any node or link types in the legend following the given value. Nodes and links in the legend are ordered
+                    by greatest to least quantity within the graph, thus, it results in the least-present types being filtered out of the legend.</p>
+                </Tab>
+                <Tab eventKey="knowledgeSources" title="Knowledge Sources">
+                  <h6>Database Sources</h6>
+                    <p>When a query is active, this setting will be populated with checkboxes for all source databases that the query was constructed from.
+                    You can then disable sources to filter them out of the graph.</p>
+                  <h6>Reasoner Sources</h6>
+                    <p>When a query is active, this setting will be populated with checkboxes for all source reasoners that the query was constructed from.
+                    You can then disable sources to filter them out of the graph.</p>
+                </Tab>
+              </Tabs>
+            )
+          },
+          {
+            title: "Table View",
+            description: `
+            This button brings up a tabular representation of the active graph. The table viewer is in a vertical split pane along with the graph,
+            so it may be resized by dragging the border where the two intersect to take up more or less space if desired.
+            Regex may also be used in the filters by entering a regex literal (e.g. /^.*$/).
+            For example, if you wanted to find all elements whose reasoner property includes RTX, you could use the pattern "/^- rtx$/m".
+            `
+          }
+        ]
+      },
+
       // Type chart
       showTypeNodes : true, // When false, shows link types (prevents far too many types being shown at once)
 
+      // Modals
+      showImportExportModal : false,
+        importForm : {
+          cacheGraph : false
+        },
+        exportForm : {
+          saveGraphState : true,
+          readable : false,
+          fileFormat: 'JSON'
+        },
+      showSettingsModal : false,
+      showTypeChart : false,
       // Cached queries modal
       cachedQueriesModalTools: [
         <FaTrash data-tip="Delete from the cache"
@@ -379,9 +662,14 @@ class App extends Component {
                    });
                  }}/>
       ],
-
-      activeModal : null,
-
+      // Help modal
+      showHelpModal : false,
+      showToolbarHelpModal : false,
+      toolbarHelpModalActiveType : 0,
+      toolbarHelpModalActiveToolType : {
+        buttons: 0,
+        tools: 0
+      },
       exampleQueries : [
           {
             title: 'Protein-Metabolite Interaction',
@@ -647,7 +935,6 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
     const untrimmedPos = codeMirror.getCursor();
     const textToCursorPositionUntrimmed = codeMirror.getRange({ line : 0, ch : 0 }, { line : pos.line, ch : pos.ch });
     const textToCursorPosition = textToCursorPositionUntrimmed.trimRight();
-    const entireText = codeMirror.getValue();
 
     // const splitLines = textToCursorPosition.split(/\r\n|\r|\n/);
     // // Adjust the position after trimming to be on the correct line.
@@ -789,9 +1076,9 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
       method: "POST",
       headers: {
         'Accept': 'application/json',
-        'Content-Type': 'application/json',
+        'Content-Type': 'text/plain',
       },
-      body: JSON.stringify([textToCursorPositionUntrimmed, entireText])
+      body: textToCursorPositionUntrimmed
     }).then(res => res.json())
       .then(async (parsedTree) => {
         setLoading(false)
@@ -816,19 +1103,12 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
             }
           }
 
-          const incompleteTree = parsedTree[0];
-          const completeTree = parsedTree[1];
-
           // Filter whitespace from the statements
-          const block = incompleteTree[incompleteTree.length-1].map((statement) => {
+          const block = parsedTree[parsedTree.length-1].map((statement) => {
             return stripLinebreaks(statement);
           });
-          const completeBlock = completeTree[completeTree.length-1].map((statement) => {
-            return stripLinebreaks(statement);
-          });
-
           const lastStatement = block[block.length-1];
-          const lastStatementComplete = completeBlock[block.length-1];
+
 
           const statementType = lastStatement[0];
 
@@ -894,22 +1174,6 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
             return predicate;
           }
 
-          const concept = (old_concept) => {
-            // Concept identifiers aren't actually parsed by the lexer, but rather the ast in Query::add.
-            // This just copies the methods that the ast uses to parse concept identifiers.
-            if (old_concept.indexOf(":") !== -1) {
-              const split = old_concept.split(":");
-              if (split.length - 1 > 1) {
-                throw new Error(`Invalid concept identifier "${old_concept}"`);
-              }
-              const [name, type_name] = split;
-              return type_name;
-            }
-            else {
-              return old_concept;
-            }
-          }
-
           const lastToken = lastStatement[lastStatement.length-1];
           const secondLastToken = lastStatement[lastStatement.length-2];
           const thirdLastToken = lastStatement[lastStatement.length-3];
@@ -938,29 +1202,23 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                 lastToken[0],
                 lastToken[1] !== undefined ? lastToken[1] : ""
               ]);
-              let previousConcept = concept(secondLastToken);
-              // May be undefined if there is no next concept
-              let nextConcept = concept(lastStatementComplete[lastStatement.length]);
-              // See https://github.com/frostyfan109/tranql/issues/117 for why this approach doesn't work
-              nextConcept = undefined;
+              let previousConcept = secondLastToken;
 
 
               const backwards = isBackwardsPredicate (currentPredicate);
 
-              console.log ([previousConcept, currentPredicate, nextConcept]);
+              console.log ([previousConcept, currentPredicate]);
 
               // Should replace this method with reduce
 
               const allEdges = graph.edges.filter((edge) => {
                 if (backwards) {
                   return edge.target_id === previousConcept &&
-                  (nextConcept === undefined || edge.source_id === nextConcept) &&
                   edge.type.startsWith(currentPredicate[1]);
                 }
                 else {
                   return (
                     edge.source_id === previousConcept &&
-                    (nextConcept === undefined || edge.target_id === nextConcept) &&
                     edge.type.startsWith(currentPredicate[1])
                   );
                 }
@@ -996,17 +1254,17 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
               }
               else if (secondLastToken === statementType) {
                 // "select foo"
-                currentConcept = concept(lastToken);
+                currentConcept = lastToken;
               }
               else if (concept_arrows.includes(lastToken) || Array.isArray(lastToken)) {
                 // "select foo->" or "select foo-[bar]->"
                 predicate = lastToken;
-                previousConcept = concept(secondLastToken);
+                previousConcept = secondLastToken;
               }
               else {
-                previousConcept = concept(thirdLastToken);
+                previousConcept = thirdLastToken;
                 predicate = secondLastToken;
-                currentConcept = concept(lastToken);
+                currentConcept = lastToken;
               }
 
 
@@ -1093,16 +1351,16 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
               let valid = true;
               if (tokens.length === 1) {
                 // Handles if there's only one concept ("select foo")
-                const currentConcept = concept(tokens[0]);
+                const currentConcept = tokens[0];
                 graph.nodes.filter((node) => node.type.startsWith(currentConcept)).forEach(node => node.reasoner.forEach((reasoner) => {
                   !validReasoners.includes(reasoner) && validReasoners.push(reasoner);
                 }));
               }
               else {
                 for (let i=0;i<tokens.length-2;i+=2) {
-                  const previousConcept = concept(tokens[i]);
+                  const previousConcept = tokens[i];
                   let predicate = tokens[i+1];
-                  const currentConcept = concept(tokens[i+2]);
+                  const currentConcept = tokens[i+2];
 
                   if (!Array.isArray(predicate)) {
                     predicate = arrowToEmptyPredicate (predicate);
@@ -1158,13 +1416,13 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
           }
           }
           catch (e) {
-            setError('Failed to parse', 'Failed to parse', [{message: e.message, details: e.stack}]);
+            setError('Failed to parse', e.message, e.stack);
           }
         }
       })
       .catch((error) => {
         if (error.name !== "AbortError") {
-          setError('Error', 'Error', [{message: error.message, details: error.stack}]);
+          setError('Error', error.message, error.stack);
         }
       });
   }
@@ -1190,7 +1448,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
         this._fgAdjustCharge (this.state.charge);
         this._updateDimensions();
       });
-      // this.setState({},()=>this._findTool.current.updateResults()); // old find tool
+      this.setState({},()=>this._findTool.current.updateResults());
     });
   }
   /**
@@ -1317,14 +1575,11 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
    */
   _analyzeAnswer (message) {
     // If we've already created the answer, use that.
-    const displayAnswerViewer = (url) => {
-      this._setActiveModal("AnswerViewerModal");
-      this._answerViewer.current.handleShow (url);
-    }
+
     if (this.state.record && this.state.record.data && this.state.record.data.hasOwnProperty ("viewURL")) {
       var url = this.state.record.data.viewURL;
       console.log ('--cached-view-url: ' + url);
-      displayAnswerViewer(url);
+      this._answerViewer.current.handleShow (url);
       //var win = window.open (url, 'answerViewer');
       //win.focus ();
       return;
@@ -1348,7 +1603,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
           message.knowledge_map = message.answers;
           delete message.answers;
           this._cacheWrite (message);
-          displayAnswerViewer(url);
+          this._answerViewer.current.handleShow (url);
           //var win = window.open (message.viewURL, 'answerViewer');
           //win.focus ();
         },
@@ -1420,12 +1675,6 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
           this._configureMessage (result[0].data);
           this._translateGraph (result[0].data);
           this.setState({ loading : false });
-          // If the query is run through the cache, then update it.
-          // this._cache.read doesn't allow us to modify it, because
-          // it converts the Dexie results into an array. Could probably?
-          // also use cacheWrite before modifying the cache result above,
-          // but this method is less prone to mysterious breakage.
-          this._cache.db["cache"].where("key").equals(result[0].key).modify({ timestamp : Date.now() });
         } else {
           // We didn't find it in the cache. Run the query.
           // Create a new controller so that the query may be aborted if desired. Abort the old one just in case.
@@ -1498,8 +1747,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
 
     var obj = {
       'key' : this.state.code,
-      'data' : cacheMessage,
-      'timestamp' : Date.now()
+      'data' : cacheMessage
     };
     // if (this.state.record) {
     //   obj.id = this.state.record.id;
@@ -1633,8 +1881,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
           NotificationManager.warning('The query returned no results', 'Warning', 4000);
         }
         let graphStateObj = isSchema ? { schema : message.graph } : { graph : message.graph };
-        // this.setState(graphStateObj, () => this._findTool.current.updateResults()); // old find tool
-        this.setState(graphStateObj);
+        this.setState(graphStateObj, () => this._findTool.current.updateResults());
       }
     });
   }
@@ -2062,6 +2309,46 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
     //   }
     // })
   }
+
+  /**
+   * Returns the formatted graph for exportation
+   *
+   * @param {Boolean} saveGraphState - Specifies whether or not the graph will save its current state
+   *
+   * @returns {Object} - The exportable graph
+   * @private
+   */
+  _getExportGraph(saveGraphState) {
+    let graph = this.state.record;
+    if (saveGraphState) {
+      graph.data.graph = GraphSerializer.serialize(this.state.graph);
+    }
+    return graph;
+  }
+  /**
+   * Dumps the graph to a given serialization format
+   *
+   * @param {Object} graph - The graph object to dump
+   * @param {String} exportType - A string of type "JSON" or "YAML"
+   * @param {Boolean} readable - A boolean indicating whether the graph should be dumped in a readable form or a minimally-sized form
+   *
+   * @returns {Object} - An object containing the dumped graph and file extension
+   * @private
+   */
+  _dumpGraph(graph,exportType,readable) {
+    let extension;
+    if (exportType === 'JSON') {
+      let indent = readable ? 2 : 0;
+      graph = JSON.stringify(graph,undefined,indent);
+      extension = '.json';
+    }
+    else if (exportType === 'YAML') {
+      let options = readable ? {} : {indent:0,noRefs:true,condenseFlow:true,noArrayIndent:true};
+      graph = YAML.safeDump(graph,options);
+      extension = '.yaml';
+    }
+    return { graph, extension };
+  }
   /**
    * Handle a click on a graph node.
    *
@@ -2126,7 +2413,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
       linkLabel: (l) => l.concatName,
       nodeRelSize:this.state.forceGraphOpts.nodeRelSize,
       enableNodeDrag:this.state.forceGraphOpts.enableNodeDrag,
-      onLinkClick:(node) => this._handleLinkClick(node, false),
+      onLinkClick:this._handleLinkClick,
       onLinkHover:this._handleLinkHover,
       onLinkRightClick:this._handleLinkRightClick,
       onNodeRightClick:this._handleNodeRightClick,
@@ -2219,13 +2506,14 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                        id="answerViewerToolbar"
                        className="App-control-toolbar fa"
                        onClick={this._handleShowAnswerViewer} />
-      <FaSearch data-tip="Find tool - helps to quickly locate specific things in the graph" id="findTool" className="App-control-toolbar fa" onClick={() => this._findTool.current.toggleShow()}/>
-      <FaQuestionCircle data-tip="Help & Information" id="helpButton" className="App-control-toolbar fa" onClick={() => this._setActiveModal('HelpModal')}/>
-      <FaDatabase data-tip="Cache Viewer - search through previous queries" id="cachedQueriesButton" className="App-control-toolbar fa" onClick={() => this._setActiveModal('CachedQueriesModal')}/>
-      <FaFolderOpen data-tip="Import/Export - Import or export graphs" id="importExportButton" className="App-control-toolbar fa" onClick={() => this._setActiveModal('ImportExportModal')}/>
-      <FaCog data-tip="Configure application settings" id="settingsToolbar" className="App-control-toolbar fa" onClick={() => this._setActiveModal('SettingsModal')} />
-      <FaTable data-active={this.state.tableViewerComponents.tableViewerCompActive} data-tip="View a tabular representation of the active graph" id="tableViewButton" className="App-control-toolbar fa" onClick={() => {
-        this.state.tableViewerComponents.tableViewerCompActive ? this._closeTableViewer() : this._openTableViewer("tableViewerCompActive");
+      <FaSearch data-tip="Find tool - helps to quickly locate specific things in the graph" id="findTool" className="App-control-toolbar fa" onClick={() => this._findTool.current.show()}/>
+      <FaQuestionCircle data-tip="Help & Information" id="helpButton" className="App-control-toolbar fa" onClick={() => this.setState({ showHelpModal : true })}/>
+      <FaDatabase data-tip="Cache Viewer - search through previous queries" id="cachedQueriesButton" className="App-control-toolbar fa" onClick={() => this._cachedQueriesModal.current.show()}/>
+      <FaFolderOpen data-tip="Import/Export - Import or export graphs" id="importExportButton" className="App-control-toolbar fa" onClick={() => this.setState({ showImportExportModal : true })}/>
+      <FaCog data-tip="Configure application settings" id="settingsToolbar" className="App-control-toolbar fa" onClick={this._handleShowModal} />
+      <FaTable data-active={this.state.tableView} data-tip="View a tabular representation of the active graph" id="tableViewButton" className="App-control-toolbar fa" onClick={() => {
+        // We want to close the table viewer if they press the button and it is already active
+        this._tableViewer.current.toggleVisibility();
       }}/>
       {
       // Perfectly functional but does not provide enough functionality as of now to warrant its presence
@@ -2278,6 +2566,14 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
       </>
     );
   }
+  /**
+   * Show the modal settings dialog.
+   *
+   * @private
+   */
+  _handleShowModal () {
+    this.setState ({ showSettingsModal : true });
+  }
   _handleShowAnswerViewer () {
     console.log (this._answerViewer);
     if (this.state.message) {
@@ -2326,6 +2622,17 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
    * @private
    */
   _handleMessageDialog (title, message, details) {
+    // Should make this a single field such as `activeModal`
+    this.setState({
+      showSettingsModal: false,
+      showTypeChart: false,
+      showHelpModal: false,
+      showToolbarHelpModal : false,
+      showImportExportModal : false
+    });
+    this._exampleQueriesModal.current.hide();
+    this._answerViewer.current.handleHide();
+
     if (!Array.isArray(message)) {
       message = [{
         message: message,
@@ -2339,8 +2646,16 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
         Otherwise, message should be an array of objects containing the properties "message" and "details".`);
     }
 
-    this._setActiveModal("ErrorModal");
     this._messageDialog.current.handleShow (title, message);
+  }
+  /**
+   * Take appropriate actions on the closing of the modal settings dialog.
+   *
+   * @private
+   */
+  _handleCloseModal () {
+    this.setState ({ showSettingsModal : false });
+    //this.setState ({ linkWeightRange : this.state.linkWeightRange});
   }
   /**
    * Handle updated settings from the modal settings dialog.
@@ -2586,43 +2901,18 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
       this.setState({ objectViewerSelection : object });
     }
   }
-  /**
-   * Closes the pane originally intended for just the table viewer
-   *
-   */
   _closeTableViewer() {
     let height = this._tableSplitPane.current.splitPane.offsetHeight;
     this._tableSplitPane.current.setState({ draggedSize : height, pane1Size : height, position : height });
     this._updateGraphSize(undefined,height);
-    const { tableViewerComponents } = this.state;
-    // Hide every component
-    Object.keys(tableViewerComponents).forEach((compKey) => {
-      tableViewerComponents[compKey] = false;
-    });
-    this.setState({ tableView : false, tableViewerComponents });
+    this.setState({ tableView : false });
   }
-  /**
-   * Opens the pane originally intended for just the table viewer
-   *
-   * @param {String} shownComponent - Refers to a boolean key of `App->state` which manages the visibility of a component rendered within the table viewer.
-   *                                   This key will be set to true.
-   *
-   * // NOTE: this entire thing is very messily written, especially the way that the state is handled--it could use a rewrite for
-   *          readability, but this works for now.
-   */
-  _openTableViewer(shownComponent) {
+  _openTableViewer() {
     const screenPortion = this.state.tableViewerSize;
     let height = this._tableSplitPane.current.splitPane.offsetHeight * screenPortion;
     this._tableSplitPane.current.setState({ draggedSize : height, pane1Size : height, position : height });
     this._updateGraphSize(undefined,height);
-    const { tableViewerComponents } = this.state;
-    // Hide every component
-    Object.keys(tableViewerComponents).forEach((compKey) => {
-      tableViewerComponents[compKey] = false;
-    });
-    // Show `shownComponent`
-    tableViewerComponents[shownComponent] = true;
-    this.setState({ tableView : true, tableViewerComponents });
+    this.setState({ tableView : true });
   }
   /**
    * Invoked on window resize
@@ -2650,6 +2940,311 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
     this._tableSplitPane.current.setState({prevWinHeight:window.innerHeight});
   }
   /**
+   * Render the help modal
+   *
+   * @private
+   */
+  _renderHelpModal () {
+    return (
+      <Modal show={this.state.showHelpModal}
+             onHide={() => this.setState({ showHelpModal : false })}
+             dialogClassName="help-modal">
+        <Modal.Header closeButton>
+          <Modal.Title>
+            Help and Information
+          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body style={{padding:"0"}}>
+          <Container id="helpGrid">
+            <Row>
+              <Col>
+                <Card>
+                  <Card.Body>
+                    <Card.Title>
+                      Documentation
+                    </Card.Title>
+                    <Card.Text>
+                      Documentation for TranQL
+                    </Card.Text>
+                    <Card.Link target="_blank" rel="noopener noreferrer" href="https://researchsoftwareinstitute.github.io/data-translator/apps/tranql">Go</Card.Link>
+                  </Card.Body>
+                </Card>
+              </Col>
+              <Col>
+                <Card>
+                  <Card.Body>
+                    <Card.Title>
+                      Examples
+                    </Card.Title>
+                    <Card.Text>
+                      Some example queries to help get you started.
+                    </Card.Text>
+                    {/*eslint-disable-next-line*/}
+                    <Card.Link href="javascript:void(0)" onClick={() => {
+                      this.setState({ showHelpModal : false });
+                      this._exampleQueriesModal.current.show();
+                    }}>View</Card.Link>
+                  </Card.Body>
+                </Card>
+              </Col>
+              <Col>
+                <Card>
+                  <Card.Body>
+                    <Card.Title>
+                      Toolbar Help
+                    </Card.Title>
+                    <Card.Text>
+                      More in-depth explanations of Toolbar's functions and what they can be used for.
+                    </Card.Text>
+                    {/*eslint-disable-next-line*/}
+                    <Card.Link href="javascript:void(0)" onClick={() => {
+                      this.setState({ showHelpModal : false, showToolbarHelpModal : true });
+                    }}>View</Card.Link>
+                  </Card.Body>
+                </Card>
+              </Col>
+            </Row>
+            <Row>
+            </Row>
+          </Container>
+        </Modal.Body>
+      </Modal>
+    );
+  }
+  /**
+   * Render the import/export modal
+   *
+   * @private
+   */
+  _renderImportExportModal () {
+    return (
+      <Modal show={this.state.showImportExportModal}
+             onHide={() => this.setState({ showImportExportModal : false })}
+             dialogClassName="import-export-modal-dialog"
+             className="import-export-modal">
+        <Modal.Header closeButton>
+          <Modal.Title>
+            Import/Export Graph
+          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <div className="no-select">
+            <div className="import-export-icon-container horizontal-bar">
+              <FaFileImport/>
+              <span>Import a graph</span>
+            </div>
+              <div className="import-options-container">
+                {<Form noValidate onSubmit={(e)=>{e.preventDefault();}} ref={this._importForm}>
+                  <Form.Check inline label="Cache the graph" name="importCacheGraph" checked={this.state.importForm.cacheGraph} onChange={(e)=>{
+                    const importForm = this.state.importForm;
+                    importForm.cacheGraph = e.target.checked;
+                    this.setState({ importForm });
+                  }}/>
+                  <FileLoader pondProps={{allowMultiple:false,maxFiles:1,acceptedFileTypes:['.json','.yaml','.yml']}}
+                              buttonProps={{type:"submit"}}
+                              filesLoadedCallback={(graph) => {
+                                graph = graph[0];
+                                if (!graph) return;
+                                const options = this.state.importForm;
+                                if (graph.hasOwnProperty('data') && graph.hasOwnProperty('key') && graph.data.hasOwnProperty('knowledge_graph')) {
+                                  this.setState({ showImportExportModal : false });
+                                  this._setSchemaViewerActive(false);
+
+                                  this._updateCode(graph.key);
+                                  this.setState({}, () => {
+                                    console.log(JSON.parse(JSON.stringify(graph)));
+                                    // this._configureMessage(graph.data);
+
+                                    let noRenderChain = false;
+                                    // If it already has a graph (save state was set to true) we should parse it so that it retains its previous state
+                                    if (graph.data.hasOwnProperty('graph')) {
+                                      noRenderChain = true;
+                                      graph.data.graph = GraphSerializer.parse(graph.data.graph);
+                                    }
+                                    this._configureMessage(graph.data,undefined,false);
+                                    this._translateGraph(graph.data, noRenderChain,false);
+                                    options.cacheGraph === true && this._cacheWrite(graph.data);
+                                  });
+
+                                }
+                                else {
+                                  this._handleMessageDialog("Graph Parsing Error", "The graph file is corrupted.", (
+                                    <div>
+                                      <p>Contains key: {graph.hasOwnProperty('key').toString()}</p>
+                                      <p>Contains knowledge_graph: {graph.hasOwnProperty('knowledge_graph').toString()}</p>
+                                      <p>Contains data: {graph.hasOwnProperty('data').toString()}</p>
+                                      <div>Object: <pre style={{display:"inline"}}>{JSON.stringify(graph,undefined,2)}</pre></div>
+                                    </div>
+                                  ));
+                                }
+                              }}
+                              loadFile={(mimeType,data) => {
+                                let message;
+                                try {
+                                  if (mimeType === "application/json") {
+                                    message = JSON.parse(data);
+                                  }
+                                  else if (mimeType === "text/yaml") {
+                                    message = YAML.safeLoad(data);
+                                  }
+                                }
+                                catch (error) {
+                                  this._handleMessageDialog("Graph Parsing Error", error.message, error.stack);
+                                }
+                                return message;
+                              }}/>
+                </Form>}
+              </div>
+          </div>
+          <div className="no-select">
+            <div className="import-export-icon-container horizontal-bar">
+              <FaFileExport/>
+              <span>Export graph{(() => {
+                if (this.state.record) {
+                  // Was lagging it so it has been removed for now
+                  // const options = this.state.exportForm;
+                  // const obj = this._getExportGraph(options.saveGraphState);
+                  // Assumes ANSI encoding
+                  // return " (" + formatBytes(this._dumpGraph(obj,options.fileFormat).graph.length,1) + ")";
+                }
+              })()}</span>
+            </div>
+              <div className="export-options-container">
+                <Form noValidate onSubmit={(e)=>{e.preventDefault();}} ref={this._exportForm}>
+                  <Form.Check inline label="Save graph state" name="exportSaveState" checked={this.state.exportForm.saveGraphState} onChange={(e)=>{
+                    const exportForm = this.state.exportForm;
+                    exportForm.saveGraphState = e.target.checked;
+                    this.setState({ exportForm });
+                  }}/>
+                  <Form.Check inline label="Export in readable form" name="exportReadable" checked={this.state.exportForm.readable} onChange={(e)=>{
+                    const exportForm = this.state.exportForm;
+                    exportForm.readable = e.target.checked;
+                    this.setState({ exportForm });
+                  }}/>
+                  <Form.Group className="form-inline">
+                    <Form.Label>File format:</Form.Label>
+                    <Form.Control as="select" name="exportFileFormat" value={this.state.exportForm.fileFormat} onChange={(e)=>{
+                      const exportForm = this.state.exportForm;
+                      exportForm.fileFormat = e.target.value;
+                      this.setState({ exportForm });
+                    }}>
+                      <option>JSON</option>
+                      <option>YAML</option>
+                    </Form.Control>
+                  </Form.Group>
+                  <div style={{width:"100%",flexGrow:1,display:"flex",justifyContent:"center",alignItems:"flex-end"}}>
+                    <Button color="primary"
+                            style={{width:"100%"}}
+                            onClick={() => {
+                              // Prevent exportation of graph if one has not been loaded
+                              // Also prevents exportation of schema (would be fairly simple to add but for now is not very useful)
+                              if (!this.state.record) {
+                                NotificationManager.warning('You must load a graph', 'Warning', 4000);
+                                return;
+                              }
+                              const options = this.state.exportForm;
+                              const exportType = options.fileFormat;
+                              const readable = options.readable;
+                              const graph = this._getExportGraph(options.saveGraphState);
+
+                              const {graph: data, extension} = this._dumpGraph(graph,exportType,readable);
+
+                              data && FileSaver.saveAs(new Blob([data]),'graph'+extension);
+                            }}
+                            {...(!this.state.record ? {className: 'disabled'} : {})}>
+                      Confirm
+                    </Button>
+                  </div>
+                </Form>
+              </div>
+          </div>
+        </Modal.Body>
+      </Modal>
+    );
+  }
+  /**
+   * Render the toolbar help modal
+   *
+   * @private
+   */
+   _renderToolbarHelpModal () {
+     const obj = {
+       buttons:React.Children.toArray(this._getButtons().props.children),
+       tools:React.Children.toArray(this._getTools().props.children)
+     };
+     return (
+       <Modal show={this.state.showToolbarHelpModal}
+              onHide={() => this.setState({ showToolbarHelpModal : false })}
+              dialogClassName="toolbar-help-modal-dialog"
+              className="toolbar-help-modal">
+         <Modal.Header closeButton>
+           <Modal.Title>
+             Help and Information
+           </Modal.Title>
+         </Modal.Header>
+         <Modal.Body>
+           <Tabs defaultActiveKey="0" className="toolbar-help-tabs">
+              {
+                Object.entries(obj).map((entry, i) => {
+                  let type = entry[0];
+                  let values = entry[1];
+
+                  const title = type.charAt(0).toUpperCase()+type.slice(1)
+                  return (
+                    <Tab eventKey={i.toString()} className="toolbar-help-tab-panel" key={i} title={title}>
+                      <ListGroup className="toolbar-help-tool-group">
+                        {
+                          values.map((val, n) => {
+                            return (
+                              // eslint-disable-next-line
+                              <ListGroup.Item className="toolbar-help-tool-button" key={n} action active={n===this.state.toolbarHelpModalActiveToolType[type]} onClick={()=>{
+                                this.state.toolbarHelpModalActiveToolType[type] = n;
+                                this.setState({ toolbarHelpModalActiveToolType : this.state.toolbarHelpModalActiveToolType })
+                              }}>
+                                {
+                                  (() => {
+                                    const noProps = (element) => {
+                                      const newProps = {};
+                                      Object.keys(element.props).forEach((k) => {
+                                        // I don't know why this is necessary but it is.
+                                        newProps[k] = undefined
+                                      });
+                                      return newProps;
+                                    }
+                                    const el = type === "tools" ? val.props.children : val;
+                                    return React.cloneElement(el, noProps(el))
+                                  })()
+                                }
+                              </ListGroup.Item>
+                            );
+                          })
+                        }
+                        <ListGroup.Item/>
+                      </ListGroup>
+                      <Card className="toolbar-help-content">
+                        <Card.Body className="toolbar-help-content-body">
+                          <Card.Header className="toolbar-help-content-title">
+                          {this.state.toolHelpDescriptions[type][this.state.toolbarHelpModalActiveToolType[type]].title}
+                          </Card.Header>
+                          <Card.Text as="div">
+                            <div>
+                              {
+                                this.state.toolHelpDescriptions[type][this.state.toolbarHelpModalActiveToolType[type]].description
+                              }
+                            </div>
+                          </Card.Text>
+                        </Card.Body>
+                      </Card>
+                    </Tab>
+                  );
+                })
+              }
+            </Tabs>
+         </Modal.Body>
+       </Modal>
+     );
+   }
+  /**
    * Render the type bar chart modal
    *
    * @private
@@ -2671,8 +3266,8 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
       }
     ));
     return (
-      <Modal show={this.state.activeModal==="TypeChartModal"}
-             onHide={() => this._setActiveModal(null)}
+      <Modal show={this.state.showTypeChart}
+             onHide={() => this.setState ({ showTypeChart : false })}
              dialogClassName="type-chart">
         <Modal.Header closeButton>
           <Modal.Title id="typeChartTitle">
@@ -2771,11 +3366,11 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
    *
    * @private
    */
-  _renderSettingsModal () {
+  _renderModal () {
     return (
       <>
-        <Modal show={this.state.activeModal==="SettingsModal"}
-               onHide={() => this._setActiveModal(null)}>
+        <Modal show={this.state.showSettingsModal}
+               onHide={this._handleCloseModal}>
           <Modal.Header closeButton>
             <Modal.Title>Settings</Modal.Title>
           </Modal.Header>
@@ -2927,37 +3522,13 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
             </Tabs>
           </Modal.Body>
           <Modal.Footer>
-            <Button variant="secondary" onClick={() => this._setActiveModal(null)}>
+            <Button variant="secondary" onClick={this._handleCloseModal}>
               Close
             </Button>
           </Modal.Footer>
         </Modal>
       </>
     );
-  }
-  /**
-   * Sets the active modal
-   *
-   * @param {string|null} modalName - Sets activeModal to modalName, if null no modal is active.
-   */
-  _setActiveModal(modalName) {
-    this.setState({ activeModal : modalName });
-  }
-  /**
-   * Handles parameters from the query string
-   *
-   */
-  _handleQueryString() {
-    // `ignoreQueryPrefix` automatically truncates the leading question mark within a query string in order to prevent it from being interpreted as part of it
-    const params = qs.parse(window.location.search, { ignoreQueryPrefix : true });
-
-    const tranqlQuery = params.q || params.query;
-    if (tranqlQuery !== undefined) {
-      this._updateCode(tranqlQuery);
-      this.setState({}, () => {
-        this._executeQuery();
-      });
-    }
   }
   /**
    * Perform any necessary cleanup before being unmounted
@@ -2980,12 +3551,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
     this._updateDimensionsFunc = this._updateDimensions;
     window.addEventListener('resize', this._updateDimensionsFunc);
 
-    // Hydrate persistent state from local storage
     this._hydrateState ();
-
-    // Handle query string parameters (mini-router implementation)
-    // & hydrate state accordingly
-    this._handleQueryString ();
 
     // Populate the cache viewer
     this._updateCacheViewer ();
@@ -3022,58 +3588,15 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
     // Render it.
     return (
       <div className="App" id="AppElement">
-        {this._renderSettingsModal () }
+        {this._renderModal () }
         {this._renderTypeChart ()}
-        <HelpModal activeModal={this.state.activeModal} setActiveModal={this._setActiveModal}/>
-        <ToolbarHelpModal activeModal={this.state.activeModal}
-                          setActiveModal={this._setActiveModal}
-                          buttons={this._getButtons()}
-                          tools={this._getTools()}
-                          />
-        <ImportExportModal activeModal={this.state.activeModal}
-                           setActiveModal={this._setActiveModal}
-                           record={this.state.record}
-                           graph={this.state.graph}
-                           code={this.state.code}
-                           importGraph={(graph, options) => {
-                             if (graph.hasOwnProperty('data') && graph.hasOwnProperty('key') && graph.data.hasOwnProperty('knowledge_graph')) {
-                               this._setActiveModal(null); // hide import/export modal
-                               this._setSchemaViewerActive(false);
-
-                               this._updateCode(graph.key);
-                               this.setState({}, () => {
-                                 console.log(JSON.parse(JSON.stringify(graph)));
-                                 // this._configureMessage(graph.data);
-
-                                 let noRenderChain = false;
-                                 // If it already has a graph (save state was set to true) we should parse it so that it retains its previous state
-                                 if (graph.data.hasOwnProperty('graph')) {
-                                   noRenderChain = true;
-                                   graph.data.graph = GraphSerializer.parse(graph.data.graph);
-                                 }
-                                 this._configureMessage(graph.data,undefined,false);
-                                 this._translateGraph(graph.data, noRenderChain,false);
-                                 options.cacheGraph === true && this._cacheWrite(graph.data);
-                               });
-
-                             }
-                             else {
-                               this._handleMessageDialog("Graph Parsing Error", "The graph file is corrupted.", (
-                                 <div>
-                                   <p>Contains key: {graph.hasOwnProperty('key').toString()}</p>
-                                   <p>Contains knowledge_graph: {graph.hasOwnProperty('knowledge_graph').toString()}</p>
-                                   <p>Contains data: {graph.hasOwnProperty('data').toString()}</p>
-                                   <div>Object: <pre style={{display:"inline"}}>{JSON.stringify(graph,undefined,2)}</pre></div>
-                                 </div>
-                               ));
-                             }
-                           }}/>
+        {this._renderHelpModal ()}
+        {this._renderToolbarHelpModal ()}
+        {this._renderImportExportModal ()}
         <NotificationContainer/>
         <QueriesModal ref={this._exampleQueriesModal}
-                      show={this.state.activeModal==="ExampleQueriesModal"}
-                      setActiveModal={this._setActiveModal}
                       runButtonCallback={(code, e) => {
-                        this._setActiveModal(null); // hide
+                        this._exampleQueriesModal.current.hide();
                         this._updateCode(code);
                         this.setState({}, () => {
                           this._executeQuery();
@@ -3083,10 +3606,8 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                       title="Example queries"/>
         <QueriesModal ref={this._cachedQueriesModal}
                       id="cachedQueriesModal"
-                      show={this.state.activeModal==="CachedQueriesModal"}
-                      setActiveModal={this._setActiveModal}
                       runButtonCallback={(code, e) => {
-                        this._setActiveModal(null); // hide
+                        this._cachedQueriesModal.current.hide();
                         this._updateCode(code);
                         this.setState({}, () => {
                           this._executeQuery();
@@ -3101,7 +3622,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
         <header className="App-header" >
           <div id="headerContainer" className="no-select">
             <p style={{display:"inline-block",flex:1}}>TranQL</p>
-            <Message activeModal={this.state.activeModal} ref={this._messageDialog} />
+            <Message show={false} ref={this._messageDialog} />
             <GridLoader
               css={spinnerStyleOverride}
               id={"spinner"}
@@ -3133,7 +3654,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
               )
             }
             <div id="appControlContainer" style={{display:(this.state.toolbarEnabled ? "none" : "")}}>
-              <FaCog data-tip="Configure application settings" id="settings" className="App-control" onClick={() => this._setActiveModal("SettingsModal")} />
+              <FaCog data-tip="Configure application settings" id="settings" className="App-control" onClick={this._handleShowModal} />
               <FaPlayCircle data-tip="Answer Navigator - see each answer, its graph structure, links, knowledge source and literature provenance" id="answerViewer" className="App-control" onClick={this._handleShowAnswerViewer} />
             </div>
           </div>
@@ -3147,10 +3668,6 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                     this.setState({ showCodeMirror : false });
                     localStorage.setItem('showCodeMirror', false);
                   }} className="editor-vis-control legend-vis-control"/>
-                  <HistoryViewer activeModal={this.state.activeModal}
-                                 setActiveModal={this._setActiveModal}
-                                 cache={this._cache}
-                                 setCode={this._updateCode}/>
                   <CodeMirror editorDidMount={(editor)=>{this._codemirror = editor;}}
                   className="query-code"
                   value={this.state.code}
@@ -3288,15 +3805,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                                           }
                                         }}/>
                         </div>
-                        <FindTool2 graph={this.state.schemaViewerActive && this.state.schemaViewerEnabled ? this.state.schema : this.state.graph}
-                                   ref={this._findTool}
-                                   resultMouseEnter={(result) => {
-                                     this._highlightType(result.id, 0xff0000, false, undefined, "id")
-                                   }}
-                                   resultMouseLeave={(result) => {
-                                     this._highlightType(result.id, false, false, undefined, "id")
-                                   }}/>
-                        {/*<FindTool graph={this.state.schemaViewerActive && this.state.schemaViewerEnabled ? this.state.schema : this.state.graph}
+                        <FindTool graph={this.state.schemaViewerActive && this.state.schemaViewerEnabled ? this.state.schema : this.state.graph}
                                   resultMouseClick={(values)=>{
                                     const isNode = function(element) {
                                       return !element.origin.hasOwnProperty('source_id') && !element.origin.hasOwnProperty('target_id');
@@ -3320,12 +3829,10 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                                   resultMouseLeave={(values)=>{
                                     values.forEach((element) => this._highlightType(element.id,false,false,undefined,'id'))}
                                   }
-                                  ref={this._findTool}/>*/}
-
+                                  ref={this._findTool}/>
                       </div>
                     </div>
-                    {/*<div onContextMenu={this._handleContextMenu} id="graphContainer" data-vis-mode={this.state.visMode}>*/}
-                    <div id="graphContainer" data-vis-mode={this.state.visMode}>
+                    <div onContextMenu={this._handleContextMenu} id="graphContainer" data-vis-mode={this.state.visMode}>
                       {this.state.schemaViewerActive && this.state.schemaViewerEnabled ?
                         (
                           this._renderForceGraph (
@@ -3342,7 +3849,7 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                           })
                         )
                       }
-                      {/*<ContextMenu id={this._contextMenuId} ref={this._contextMenu}/>*/}
+                      <ContextMenu id={this._contextMenuId} ref={this._contextMenu}/>
                     </div>
                   </div>
                   <div id="info" style={!this.state.objectViewerEnabled ? {display:"none"} : {}}>
@@ -3366,9 +3873,8 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                 </SplitPane>
               }
             </div>
-            <div className="h-100">
-            <TableViewer tableView={this.state.tableViewerComponents.tableViewerCompActive}
-                         close={this._closeTableViewer}
+            <TableViewer onOpen={this._openTableViewer}
+                         onClose={this._closeTableViewer}
                          ref={this._tableViewer}
                          data={(() => {
                            const graph = this.state.schemaViewerActive && this.state.schemaViewerEnabled ? this.state.schema : this.state.graph;
@@ -3418,7 +3924,6 @@ SELECT population_of_individual_organisms->chemical_substance->gene->biological_
                              };
                            }
                          }}/>
-            </div>
           </SplitPane>
         </div>
         <div id='next'/>
